requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
jinja2==3.1.4
python-multipart==0.0.9
sqlalchemy==2.0.31
pypdf==4.2.0
pillow==10.4.0
pytesseract==0.3.10
python-dateutil==2.9.0.post0

openai>=1.0.0
